% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.info.hg38}
\alias{gene.info.hg38}
\title{gene.info.hg38}
\format{
An object of class \code{data.frame} with 59609 rows and 15 columns.
}
\usage{
gene.info.hg38
}
\description{
Gene annotation table for hg38 (Ensembl release 101)
}
\examples{
head(gene.info.hg38)
}
\keyword{datasets}
