% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.info.mm10}
\alias{gene.info.mm10}
\title{gene.info.mm10}
\format{
An object of class \code{data.frame} with 55401 rows and 17 columns.
}
\usage{
gene.info.mm10
}
\description{
Gene annotation table for mm10 (Ensembl release 101)
}
\examples{
head(gene.info.mm10)
}
\keyword{datasets}
