% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBedOverlap.R
\name{myBedOverlap}
\alias{myBedOverlap}
\title{Venn plot show intersection of 2 Granges}
\usage{
myBedOverlap(
  bed1,
  bed2,
  name1 = "bed1",
  name2 = "bed2",
  extend = 0,
  ignore.strand = TRUE
)
}
\arguments{
\item{bed1}{First Grange object}

\item{bed2}{Second Grange object}

\item{name1}{Name of the first Grange object}

\item{name2}{Name of the second Grange object}

\item{extend}{Number of bases to extend in each Grange}

\item{ignore.strand}{Ignore strand or not}
}
\value{
Venn object of bed overlap
}
\description{
Venn plot show intersection of 2 Granges
}
\examples{
myBedOverlap(GM12878.H3K4me3.rep1, GM12878.H3K4me3.rep2, extend = 50)
}
