% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFilterCNV.R
\name{myFilterCNV}
\alias{myFilterCNV}
\title{Filter single-cell CNV by region length}
\usage{
myFilterCNV(segcopy, nbins = 10)
}
\arguments{
\item{segcopy}{Cell-by-bin matrix of copy number}

\item{nbins}{Cutoff for number of continious bins with CNV}
}
\value{
Filtered cell-by-bin matrix of copy number
}
\description{
Filter single-cell CNV by region length
}
\examples{
myFilterCNV(segcopy, nbins = 10)
}
