% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGeneSensitivity.R
\name{myGeneSensitivity}
\alias{myGeneSensitivity}
\title{Summarize sensitivity from scRNA-seq read/UMI count matrix}
\usage{
myGeneSensitivity(rc, n = 20, cutoff = 1)
}
\arguments{
\item{rc}{Read count matrix}

\item{n}{Number of quantile bins to break into}

\item{cutoff}{Minimal number of reads/UMIs for detection}
}
\value{
Data.frame with quantile breaks, accululated genes and standard deviation among cells
}
\description{
Summarize sensitivity from scRNA-seq read/UMI count matrix
}
\examples{
myGeneSensitivity(rc, 20, 1)
}
