% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGetDensity.R
\name{myGetDensity}
\alias{myGetDensity}
\title{Get density of scatter plot by 2D KDE}
\usage{
myGetDensity(x, y, nbins = 100, ...)
}
\arguments{
\item{x}{X coordination}

\item{y}{Y coordination}

\item{nbins}{Number of bins to break into}

\item{...}{other paramerters passed to kede2d}
}
\value{
Density for each dot
}
\description{
Get density of scatter plot by 2D KDE
}
\examples{
myGetDensity(plot.data$x, plot.data$y)
}
