% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLogNorm.R
\name{myLogNorm}
\alias{myLogNorm}
\title{Normalize read/UMI count matrix}
\usage{
myLogNorm(read.count, scale.factor = 10000, log = TRUE)
}
\arguments{
\item{read.count}{Read count/UMI matrix}

\item{scale.factor}{Scale factor}

\item{log}{Perform log transformation or not}
}
\value{
Normalized count matrix
}
\description{
Normalize read/UMI count matrix
}
\examples{
myLogNorm(rc, 1e6, TRUE)
}
