% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMultiPlot.R
\name{myMultiPlot}
\alias{myMultiPlot}
\title{Arrange multiple ggplot in grid}
\usage{
myMultiPlot(..., plotlist = NULL, file, cols = 1, layout = "row")
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{List of ggplot objects}

\item{file}{Output file}

\item{cols}{Number of columns}

\item{layout}{Order by "row" or "col"}
}
\value{
plots
}
\description{
Arrange multiple ggplot in grid
}
\examples{
myMultiPlot(p1, p2, p3, p4, cols = 2)
myMultiPlot(plotlist = p.list, cols = 2)
}
