% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotCNProfile.R
\name{myPlotCNProfile}
\alias{myPlotCNProfile}
\title{Plot single-cell copy number dot plot}
\usage{
myPlotCNProfile(
  segref,
  segcopy,
  bininfo,
  sex = "XX",
  cell = "Cell",
  top = 8,
  return = FALSE
)
}
\arguments{
\item{segref}{Number of normalized read count of genomic bin}

\item{segcopy}{Copy number of genomic bin}

\item{bininfo}{Metadata for genomic bin, bininfo$CHR indicates chromosome}

\item{sex}{"XX" for female and "XY" for male}

\item{cell}{Name of the single cell}

\item{top}{Maximum copy number to be plotted}

\item{return}{Reture plot or not}
}
\value{
ggplot object of single cell CN profile
}
\description{
Plot single-cell copy number dot plot
}
\examples{
myPlotCNProfile(segref$sc1, segcopy$sc1, bininfo, cell = "sc1")
}
