% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPrintPairwiseAlignment.R
\name{myPrintPairwiseAlignment}
\alias{myPrintPairwiseAlignment}
\title{Print pairwise alignment of pairwiseAlignment}
\usage{
myPrintPairwiseAlignment(alignment, chunksize = 60)
}
\arguments{
\item{alignment}{pairwiseAlignment object}

\item{chunksize}{bases per line}
}
\value{
list of pairwiseAlignment
}
\description{
Print pairwise alignment of pairwiseAlignment
}
\examples{
myPrintPairwiseAlignment(pwAln)
}
