% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRevSequence.R
\name{myRevSequence}
\alias{myRevSequence}
\title{Reverse complement of DNA sequence}
\usage{
myRevSequence(text, complementary = TRUE)
}
\arguments{
\item{text}{DNA sequence with ATCG}

\item{complementary}{Whether perform complement or not}
}
\value{
Reversed DNA sequence
}
\description{
Reverse complement of DNA sequence
}
\examples{
myRevSequence("ATGATCGTTCA", complementary = TRUE)
}
