% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protien.info.hg38}
\alias{protien.info.hg38}
\title{protien.info.hg38}
\format{
An object of class \code{data.frame} with 112012 rows and 27 columns.
}
\usage{
protien.info.hg38
}
\description{
Protein annotation table for hg38 (Ensembl release 101)
}
\examples{
head(protien.info.hg38)
}
\keyword{datasets}
