% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans.info.hg38}
\alias{trans.info.hg38}
\title{trans.info.hg38}
\format{
An object of class \code{data.frame} with 251298 rows and 16 columns.
}
\usage{
trans.info.hg38
}
\description{
Transcript annotation table for hg38 (Ensembl release 101)
}
\examples{
head(trans.info.hg38)
}
\keyword{datasets}
